Add Export Scenarios functionality.
Closes #29

My longer term goal is to allow a project to include different
resourcing scenarios. Ultimately, this will also require the ability
to include changes in specific resource allocation to activities.

In the near term, my approach was to auto-generate scenarios by taking
every non-explicit resource and generate scenarios for the minimum
number of non-explicit resources (one) up to the total amount of
non-explicit resources. For example, in a project configured for a
 maximum of 6 devs who are all non-explicit resource targets, six
 scenarios would be generated, each with 1, 2, 3, 4, 5, and finally
 6 devs respectively.

The new Export Scenarios menu item in the File menu will generate a
csv file that contains the calculated metrics for each scenario (Costs,
Criticality Metrics, Cyclomatic Complexity and Duration Man Months).

This could also form a foundational step toward providing a new graph
tab that displays the scenario metrics in comparison to one another,
greatly simplifying the ability to discover the area of minimum cost.